requirement.txt
torch
torchvision
flwr
pillow
matplotlib
seaborn
pandas
numpy
scipy
scikit-posthocs
